+++ old/publicOld/srcOld/package.json
@@ -1,20 +1,21 @@
 {
   "name": "src",
   "version": "1.0.0",
-  "description": "",
-  "main": "client.js",
+  "main": "server.js",
   "scripts": {
     "devStart": "nodemon server.js",
     "start": "node server.js"
   },
   "keywords": [],
-  "author": "",
+  "author": "Maki Spariva",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.2",
     "cors": "^2.8.5",
+    "express": "^4.18.2"
+  },
+  "devDependencies": {
     "dotenv": "^16.4.1",
-    "express": "^4.18.2",
     "nodemon": "^3.0.3"
-  }
+  },
+  "description": ""
 }
