--- src/gt sergio/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "gt-sergio",
-  "version": "1.0.0",
-  "description": "",
-  "main": "express.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.18.2"
-  }
-}
