--- src/gt sergio/tareas.json
@@ -1,1 +1,0 @@
-[{"id":1706674419790,"nombre":"adszdz","categoria":"yug","fecha":"2024-01-03","color":"naranja"},{"id":1706674474541,"nombre":"adszdz","categoria":"yug","fecha":"2024-01-03","color":"naranja"}]
