+++ src/tareas.json
@@ -1,0 +1,1 @@
+[{"id":1706674474541,"nombre":"adszdz","categoria":"yug","fecha":"2024-01-03","color":"naranja"},{"id":1706720644638,"nombre":"SFDA","categoria":"SAFD","fecha":"2024-01-04","color":"rojo"},{"id":1706720661598,"nombre":"ADF","categoria":"AFDS","fecha":"","color":"negro"},{"id":1706720665400,"nombre":"AFD","categoria":"AFDS","fecha":"","color":"negro"},{"id":1706720672557,"nombre":"FASFS","categoria":"FASFA","fecha":"","color":"negro"},{"id":1706720675933,"nombre":"FAFS","categoria":"AFDFSA","fecha":"","color":"negro"}]
